+++ src/ChessPackage/Pieces/Queen.java
@@ -27,10 +27,10 @@ public class Queen extends Piece {
         // Disallow 'moving' to the startSquare square or to a square with a friendly piece
         if (startSquare.equals(endSquare) || (endSquare.getPiece() != null && this.color == endSquare.getPiece().getColor()))
             return false;
-        byte startFile = startSquare.getFile();
-        byte startRank = startSquare.getRank();
-        byte endFile = endSquare.getFile();
-        byte endRank = endSquare.getRank();
+        int startFile = startSquare.getFile();
+        int startRank = startSquare.getRank();
+        int endFile = endSquare.getFile();
+        int endRank = endSquare.getRank();
         // Check for Rook-like moves
         // Going up the board, check if there are any pieces in the way
         if (startFile == endFile && endRank > startRank)
